refactor: extract utilities and add library structure
